+++ CHANGELOG.md
@@ -151,6 +151,8 @@ In order to build the DUNE core modules you need at least the following software
   `DUNE_ASSERT_AND_RETURN`. Use `assert()` macro directly in `constexpr` functions.
 - Remove deprecated header `power.hh`. Use `Dune::power` from `math.hh` instead.
 - Deprecate class `SizeOf`. Use `sizeof...` instead.
+- Deprecate header `dune/common/keywords.hh` and the provided macros
+  `DUNE_GENERALIZED_CONSTEXPR` and `DUNE_INLINE_VARIABLE`. Use the key words directly.
 - Remove deprecated header `dune/python/common/numpycommdatahandle.hh`. Use
   `dune/python/grid/numpycommdatahandle.hh` instead.